pyproject.toml
# Copyright (c) 2025-2026 Datalayer, Inc.
#
# BSD 3-Clause License

[build-system]
requires = ["hatchling~=1.21"]
build-backend = "hatchling.build"

[project]
name = "codeai"
authors = [{ name = "Datalayer", email = "info@datalayer.ai" }]
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["AI", "AI Agents", "Code Execution", "Jupyter", "CLI"]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
]
dependencies = [
    "agent-runtimes",
    "httpx",
    "jupyter-kernel-client",
    "jupyter-nbmodel-client",
    "pillow",
    "prompt_toolkit>=3.0.0",
    "requests",
    "rich>=13.0.0",
    "textual",
    "typer>=0.9.0",
]

[project.optional-dependencies]
test = ["ipykernel", "jupyter_server>=1.6,<3", "pytest>=7.0"]
lint = ["mdformat>0.7", "mdformat-gfm>=0.3.5", "ruff"]
typing = ["mypy>=0.990"]

[project.license]
file = "LICENSE"

[project.urls]
Home = "https://github.com/datalayer/codeai"

[project.scripts]
codeai = "codeai.cli:main"

[tool.hatch.version]
path = "codeai/__version__.py"

[tool.pytest.ini_options]
filterwarnings = [
  "error",
  "ignore:There is no current event loop:DeprecationWarning",
  "module:make_current is deprecated:DeprecationWarning",
  "module:clear_current is deprecated:DeprecationWarning",
  "module:Jupyter is migrating its paths to use standard platformdirs:DeprecationWarning",
]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
pretty = true
show_error_context = true
show_error_codeai = true
strict_equality = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
  "A",
  "B",
  "C",
  "E",
  "F",
  "FBT",
  "I",
  "N",
  "Q",
  "RUF",
  "S",
  "T",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # FBT001 Boolean positional arg in function definition
  "FBT001",
  "FBT002",
  "FBT003",
]

[tool.ruff.lint.per-file-ignores]
# S101 Use of `assert` detected
"codeai/tests/*" = ["S101"]